[Mod] Weapon의 데이터를 불러와 적용시킬 AttackStatus 추가 및 적용

+++ Assets/Scripts/Weapon/Weapon.cs
@@ -34,7 +34,7 @@ public class WeaponStat
             this.ShootingModes = ShootingModes;
         }
 
-        _fireRate = 1 / (this.RoundsPerMinute / 60);
+        _fireRate = 1 / (float)(this.RoundsPerMinute / 60);
         FireRateWFS = YieldCache.WaitForSeconds(_fireRate);
     }
 }
